Refactor Downloader as a Helper class
